chore(refactor): move modules around, get rid of src/services subdir
